% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_baadviz.R
\name{run_baadviz}
\alias{run_baadviz}
\title{Visualize the Biomass And Allometry Database (BAAD)
You must have the baad.data package installed. Go here to install it:
https://github.com/traitecoevo/baad.data}
\usage{
run_baadviz()
}
\description{
Visualize the Biomass And Allometry Database (BAAD)
You must have the baad.data package installed. Go here to install it:
https://github.com/traitecoevo/baad.data
}

